long_shortmsg protobuf log

+++ smsapp/CnppCodeAndDecode.cpp
@@ -2162,7 +2162,7 @@ int CCnppCodeAndDecode::fnPutMessageHeader(char **cppDestBuff,int& nDestSize,cha
 	nEN=fnPutToCode(cppDestBuff,nDestSize,pMH->Command_Length,sizeof(pMH->Command_Length));
 	if(nEN!=CODE_SUCCEED) 
 		return(nEN);
-	nEN=fnPutToCode(cppDestBuff,nDestSize,pMH->Command_ID,sizeof(pMH->Command_ID));
+	nEN=fnPutToCode(cppDestBuff,nDestSize,(long)pMH->Command_ID,sizeof(pMH->Command_ID));
 	if(nEN!=CODE_SUCCEED) 
 		return(nEN);
 	// modify by lzh
